requirements.txt
python-gitlab
cryptography >= 1.4.0
GitPython